% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_panel.R
\name{assay_panel_server}
\alias{assay_panel_server}
\title{Assay Panel Server}
\usage{
assay_panel_server(assay_name, obj, output, input)
}
\arguments{
\item{assay_name}{Name of the assay}

\item{obj}{A Seurat object}

\item{output}{Shiny output object}

\item{input}{Shiny input object}
}
\value{
NULL (called for side effects)
}
\description{
Server logic for rendering assay data.
}
\keyword{internal}
