% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modules.R
\name{assay_panel_ui}
\alias{assay_panel_ui}
\title{Assay Panel UI}
\usage{
assay_panel_ui(assay_name, obj)
}
\arguments{
\item{assay_name}{Name of the assay}

\item{obj}{A Seurat object}
}
\value{
A shiny fluidRow UI element
}
\description{
Generate the UI for a single assay panel.
}
\keyword{internal}
