% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_matrix_heatmap}
\alias{create_matrix_heatmap}
\title{Create Matrix Heatmap}
\usage{
create_matrix_heatmap(matrix_data, title = "Matrix Heatmap")
}
\arguments{
\item{matrix_data}{A matrix to plot}

\item{title}{Plot title}
}
\value{
A plotly object
}
\description{
Create an interactive heatmap of matrix data.
}
\keyword{internal}
