% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{extract_embeddings}
\alias{extract_embeddings}
\title{Extract Embeddings}
\usage{
extract_embeddings(obj, reduction_name, dims = c(1, 2))
}
\arguments{
\item{obj}{A Seurat object}

\item{reduction_name}{Name of the reduction}

\item{dims}{Vector of two dimensions to extract}
}
\value{
A data frame with cell names and embedding coordinates
}
\description{
Extract dimensional reduction embeddings from a Seurat object.
}
\keyword{internal}
