% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{FindIdentLabel}
\alias{FindIdentLabel}
\title{Find Ident Label}
\usage{
FindIdentLabel(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
The name of the matching metadata column, or "Unknown"
}
\description{
Find the metadata column name that matches the active Idents.
}
\keyword{internal}
