% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{get_assay_data_safe}
\alias{get_assay_data_safe}
\title{Get Assay Data Safely}
\usage{
get_assay_data_safe(obj, assay_name, layer)
}
\arguments{
\item{obj}{A Seurat object}

\item{assay_name}{Name of the assay}

\item{layer}{Name of the layer (counts, data, scale.data)}
}
\value{
Matrix data or NULL if not available
}
\description{
Wrapper around GetAssayData with error handling.
}
\keyword{internal}
