% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_panel.R
\name{get_assay_info}
\alias{get_assay_info}
\title{Get Assay Information}
\usage{
get_assay_info(obj, assay_name)
}
\arguments{
\item{obj}{A Seurat object}

\item{assay_name}{Name of the assay to get information for}
}
\value{
A list with assay metadata including dimensions and available layers
}
\description{
Extract information about a specific assay from a Seurat object.
}
\keyword{internal}
