% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{get_matrix_sample}
\alias{get_matrix_sample}
\title{Get Matrix Sample}
\usage{
get_matrix_sample(matrix, max_rows = 10, max_cols = 20)
}
\arguments{
\item{matrix}{A matrix (dense or sparse)}

\item{max_rows}{Maximum number of rows to return}

\item{max_cols}{Maximum number of columns to return}
}
\value{
A dense matrix sample
}
\description{
Extract a sample of rows and columns from a matrix for display.
}
\keyword{internal}
