% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{get_sparsity_info}
\alias{get_sparsity_info}
\title{Get Sparsity Information}
\usage{
get_sparsity_info(matrix)
}
\arguments{
\item{matrix}{A matrix (dense or sparse)}
}
\value{
A list with sparsity statistics
}
\description{
Calculate sparsity statistics for a matrix.
}
\keyword{internal}
