% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{get_top_variable_features}
\alias{get_top_variable_features}
\title{Get Top Variable Features}
\usage{
get_top_variable_features(obj, assay_name, n = 20)
}
\arguments{
\item{obj}{A Seurat object}

\item{assay_name}{Name of the assay}

\item{n}{Number of features to return}
}
\value{
Character vector of feature names
}
\description{
Get the top variable features for an assay.
}
\keyword{internal}
