% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_feature_expression}
\alias{plot_feature_expression}
\title{Plot Feature Expression}
\usage{
plot_feature_expression(obj, reduction_name, feature, dims = c(1, 2))
}
\arguments{
\item{obj}{A Seurat object}

\item{reduction_name}{Name of the reduction to plot}

\item{feature}{Gene/feature name to plot}

\item{dims}{Vector of two dimensions to plot}
}
\value{
A ggplot object
}
\description{
Create a scatter plot showing expression of a single feature.
}
\keyword{internal}
