% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_feature_heatmap}
\alias{plot_feature_heatmap}
\title{Plot Feature Heatmap}
\usage{
plot_feature_heatmap(obj, assay_name, features = NULL, n_cells = 100, config)
}
\arguments{
\item{obj}{A Seurat object}

\item{assay_name}{Name of the assay}

\item{features}{Optional character vector of features to plot}

\item{n_cells}{Number of cells to sample}

\item{config}{Configuration list with plot settings}
}
\value{
A ggplot object or NULL
}
\description{
Create a heatmap of top variable features.
}
\keyword{internal}
