% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_metadata_distribution}
\alias{plot_metadata_distribution}
\title{Plot Metadata Distribution}
\usage{
plot_metadata_distribution(obj, column, config)
}
\arguments{
\item{obj}{A Seurat object}

\item{column}{Metadata column name}

\item{config}{Configuration list with plot settings}
}
\value{
A ggplot object
}
\description{
Create a histogram or bar plot showing distribution of a metadata column.
}
\keyword{internal}
