% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_qc_metrics}
\alias{plot_qc_metrics}
\title{Plot QC Metrics}
\usage{
plot_qc_metrics(obj, config)
}
\arguments{
\item{obj}{A Seurat object}

\item{config}{Configuration list with plot settings}
}
\value{
A list of ggplot objects or NULL
}
\description{
Create histograms of common QC metrics.
}
\keyword{internal}
