% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_reduction_interactive}
\alias{plot_reduction_interactive}
\title{Plot Reduction Interactive}
\usage{
plot_reduction_interactive(
  obj,
  reduction_name,
  color_by = "None",
  dims = c(1, 2),
  config
)
}
\arguments{
\item{obj}{A Seurat object}

\item{reduction_name}{Name of the reduction to plot}

\item{color_by}{Metadata column to color by, or "None"}

\item{dims}{Vector of two dimensions to plot}

\item{config}{Configuration list with plot settings}
}
\value{
A plotly object
}
\description{
Create an interactive plotly scatter plot for dimensional reduction.
}
\keyword{internal}
