% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_reduction_static}
\alias{plot_reduction_static}
\title{Plot Reduction Static}
\usage{
plot_reduction_static(obj, reduction_name, color_by = "None", dims = c(1, 2))
}
\arguments{
\item{obj}{A Seurat object}

\item{reduction_name}{Name of the reduction to plot}

\item{color_by}{Metadata column to color by, or "None"}

\item{dims}{Vector of two dimensions to plot}
}
\value{
A ggplot object
}
\description{
Create a static ggplot scatter plot for dimensional reduction.
Faster than interactive version for large datasets.
}
\keyword{internal}
