% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{SeuratInfo}
\alias{SeuratInfo}
\title{Get Seurat Object Information}
\usage{
SeuratInfo(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
A list with formatted information about the object
}
\description{
Generate a comprehensive summary of a Seurat object.
}
