% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{validate_seurat_object}
\alias{validate_seurat_object}
\title{Validate Seurat Object}
\usage{
validate_seurat_object(obj)
}
\arguments{
\item{obj}{An object to validate}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Check that a Seurat object is valid for viewing.
}
\keyword{internal}
