% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{viewseurat-app}
\alias{viewseurat-app}
\title{ViewSeurat Shiny Application Components}
\description{
These functions create the UI and server components for the ViewSeurat Shiny app.
}
\keyword{internal}
