% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{viewseurat_app}
\alias{viewseurat_app}
\title{Create the ViewSeurat Shiny Application}
\usage{
viewseurat_app()
}
\value{
A Shiny app object
}
\description{
This function creates and returns the Shiny app object for viewing Seurat objects.
It is called internally by \code{\link{ViewSeurat}}.
}
\keyword{internal}
