% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{viewseurat_server}
\alias{viewseurat_server}
\title{Create the ViewSeurat Server Function}
\usage{
viewseurat_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
Create the ViewSeurat Server Function
}
\keyword{internal}
